requirements.txt
# Framework Web
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2

# Banco de Dados PostgreSQL
psycopg2-binary==2.9.9

# Variáveis de Ambiente
python-dotenv==1.0.0

# Geração de PDF
reportlab==4.0.7

# Planilhas Excel
openpyxl==3.1.2
et-xmlfile==1.1.0

# Markdown (para formatação de atas)
markdown==3.5.1

# Agendamento de tarefas (opcional)
schedule==1.2.0

# Autenticação e Segurança
bcrypt==4.0.1
cryptography==41.0.7

# Utilitários
python-dateutil==2.8.2
pytz==2023.3